Split character builder runtime hooks
